ebdjango/requirements.txt
asgiref==3.7.2
boto3==1.26.142
botocore==1.29.142
distlib==0.3.6
Django==3.2.19
django-cors-headers==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
filelock==3.12.0
jmespath==1.0.1
platformdirs==3.5.1
PyJWT==2.7.0
python-dateutil==2.8.2
pytz==2023.3
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
urllib3==1.26.16
virtualenv==20.23.0
drf-yasg